The unittest package's workhorse function

Report the test of an expression in TAP format.

ok(test, description)

Arguments

test

Expression to be tested. Evaluating to TRUE is treated as success, anything else as failure.

description

Character string describing the test. If a description is not given a character representation of the test expression will be used.

Value

ok() returns whatever was returned when test is evaluated. More importantly it has the side effect of printing the result of the test in TAP format.

Details

See unittest package documentation.

The unittest.output option tells unittest where output should be sent. This is most useful for vignettes, where sending output to stderr separates the unittest output from the vignette itself.

Examples

  ok(1==1, "1 equals 1")
#> ok - 1 equals 1
  # ok - 1 equals 1

  ok(1==1)
#> ok - 1 == 1
  # ok - 1 == 1

  ok(1==2, "1 equals 2")
#> not ok - 1 equals 2
#> # Test returned non-TRUE value:
#> # [1] FALSE
  # not ok - 1 equals 2
  # # Test returned non-TRUE value:
  # # [1] FALSE

  ok(all.equal(c(1,2),c(1,2)), "compare vectors")
#> ok - compare vectors
  # ok - compare vectors

  fn <- function () stop("oops")
  ok(fn(), "something with a coding error")
#> not ok - something with a coding error
#> # Test resulted in error:
#> #  oops
#> # Whilst evaluating:
#> #  fn()
#> # Stacktrace:
#> # -> .f(.x[[i]], ...)
#> # -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env, 
#> #     run_dont_run = run_dont_run), error = function(err) {
#> #     msg <- c(paste0("Failed to parse Rd in ", topic$file_in), 
#> #         i = err$message)
#> #     abort(msg, parent = err)
#> # })
#> # -> data_reference_topic(topic, pkg, examples_env = examples_env, 
#> #     run_dont_run = run_dont_run)
#> # -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env), 
#> #     topic = tools::file_path_sans_ext(topic$file_in), run_dont_run = run_dont_run)
#> # -> highlight_examples(code, topic, env = env)
#> # -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, 
#> #     env = child_env(env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> # -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> # -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, 
#> #     debug = debug, last = i == length(out), use_try = stop_on_error != 
#> #         2L, keep_warning = keep_warning, keep_message = keep_message, 
#> #     output_handler = output_handler, include_timing = include_timing)
#> # -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr, 
#> #     envir, enclos, user_handlers)), warning = wHandler, error = eHandler, 
#> #     message = mHandler)))
#> # -> handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr, 
#> #     envir, enclos, user_handlers)), warning = wHandler, error = eHandler, 
#> #     message = mHandler))
#> # -> try(f, silent = TRUE)
#> # -> tryCatch(expr, error = function(e) {
#> #     call <- conditionCall(e)
#> #     if (!is.null(call)) {
#> #         if (identical(call[[1L]], quote(doTryCatch))) 
#> #             call <- sys.call(-4L)
#> #         dcall <- deparse(call, nlines = 1L)
#> #         prefix <- paste("Error in", dcall, ": ")
#> #         LONG <- 75L
#> #         sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#> #         w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#> #         if (is.na(w)) 
#> #             w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], 
#> #                 type = "b")
#> #         if (w > LONG) 
#> #             prefix <- paste0(prefix, "\n  ")
#> #     }
#> #     else prefix <- "Error : "
#> #     msg <- paste0(prefix, conditionMessage(e), "\n")
#> #     .Internal(seterrmessage(msg[1L]))
#> #     if (!silent && isTRUE(getOption("show.error.messages"))) {
#> #         cat(msg, file = outFile)
#> #         .Internal(printDeferredWarnings())
#> #     }
#> #     invisible(structure(msg, class = "try-error", condition = e))
#> # })
#> # -> tryCatchList(expr, classes, parentenv, handlers)
#> # -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> # -> doTryCatch(return(expr), name, parentenv, handler)
#> # -> withCallingHandlers(withVisible(eval_with_user_handlers(expr, 
#> #     envir, enclos, user_handlers)), warning = wHandler, error = eHandler, 
#> #     message = mHandler)
#> # -> withVisible(eval_with_user_handlers(expr, envir, enclos, user_handlers))
#> # -> eval_with_user_handlers(expr, envir, enclos, user_handlers)
#> # -> eval(expr, envir, enclos)
#> # -> eval(expr, envir, enclos)
#> # -> ok(fn(), "something with a coding error")
#> # -> tryCatch(withCallingHandlers(test, error = capture_calls), error = function(e) e)
#> # -> tryCatchList(expr, classes, parentenv, handlers)
#> # -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> # -> doTryCatch(return(expr), name, parentenv, handler)
#> # -> withCallingHandlers(test, error = capture_calls)
#> # -> fn()
#> # -> stop("oops")
  # not ok - something with a coding error
  # # Test resulted in error:
  # #  oops
  # # Whilst evaluating:
  # #  fn()

  ok(c("Some diagnostic", "messages"), "A failure with diagnostic messages")
#> not ok - A failure with diagnostic messages
#> # Test returned non-TRUE value:
#> # Some diagnostic
#> # messages
  # not ok - A failure with diagnostic messages
  # # Test returned non-TRUE value:
  # # Some diagnostic
  # # messages

  ## Send unittest output to stderr()
  options(unittest.output = stderr())
  ok(ut_cmp_equal(4, 5), "4 == 5? Probably not")

  ## Reset unittest output to default (stdout())
  options(unittest.output = NULL)
  ok(ut_cmp_equal(4, 5), "4 == 5? Probably not")
#> not ok - 4 == 5? Probably not
#> # Test returned non-TRUE value:
#> # Mean relative difference: 0.25
#> # --- 4
#> # +++ 5
#> # [1] [-4-]{+5+}

  # \dontshow{
    # Clear unittest result log, so our unittest failues don't fail example-building
    rm('outcomes', pos = unittest:::pkg_vars)
    rm('errors', pos = unittest:::pkg_vars)
#> Warning: object 'errors' not found
  # }