Compare variables with verbose error output

A wrapper for all.equal and identical that provides more useful diagnostics when used in a unittest ok function.

ut_cmp_equal(a, b, filter = NULL, deparse_frame = -1, ...)
  ut_cmp_identical(a, b, filter = NULL, deparse_frame = -1)

Arguments

a

First item to compare, usually the result of whatever you are testing

b

Second item to compare, usually the expected output of whatever you are testing

filter

An optional filter function, that turns either a or b into text, and prints this out

deparse_frame

Tell sys.call which frame to deparse to get original expressions. Set to -2 when making a helper function, see examples.

...

Other arguments passed directly to all.equal

Details

For both functions, a and b are first passed to all.equal (for ut_cmp_equal()) or identical (for ut_cmp_identical()). If they match, then the function returns TRUE and your test passes.

If this fails, then we turn both a and b into text, and then use git diff to compare the 2 outputs. If you do not have git installed, then the 2 outputs will be shown side-by-side.

When using git diff, we turn colored output on when outputting to a terminal. You can force this on or off using options("cli.num_colors" = 1) or the NO_COLOR or R_CLI_NUM_COLORS environment variable.

The step of turning into text is done with the filter function. There are several of these built-in, and it will choose the one that produces the simplest output. This may mean that the output will be from the print function if the differences are obvious, or str with many decimal places if there are subtle differences between the 2.

You can also provide your own filter function if there's a particular way you would like to see the data when comparing, for example you can use write.table if your data is easiest to understand in tabular output.

Value

Returns TRUE if a & b are all.equal (for ut_cmp_equal()) or identical (for ut_cmp_identical()). Otherwise, returns an invisible() character vector of diagnostic strings helping you find where the difference is.

If called directly in an interactive R session, this output will be printed to the console.

Examples

  ## A function to test:
  fn <- function(x) { seq(x) }

  ## Get it right, and test passes:
  ok(ut_cmp_equal(fn(3), c(1,2,3)))
#> ok - ut_cmp_equal(fn(3), c(1, 2, 3))
  # ok - ut_cmp_equal(fn(3), c(1, 2, 3))

  ## Get it wrong, and we get told where in the output things are different:
  ok(ut_cmp_equal(fn(3), c(1,4,3)))
#> not ok - ut_cmp_equal(fn(3), c(1, 4, 3))
#> # Test returned non-TRUE value:
#> # Mean relative difference: 1
#> # --- fn(3)
#> # +++ c(1, 4, 3)
#> # [1] 1 [-2-]{+4+} 3
  # not ok - ut_cmp_equal(fn(3), c(1, 4, 3))
  # # Test returned non-TRUE value:
  # # Mean relative difference: 1
  # # --- fn(3)
  # # +++ c(1, 4, 3)
  # # [1] 1 [-2-]{+4+} 3

  ## Using a custom filter, we can format the output with write.table:
  ok(ut_cmp_equal(fn(3), c(1,4,3), filter = write.table))
#> not ok - ut_cmp_equal(fn(3), c(1, 4, 3), filter = write.table)
#> # Test returned non-TRUE value:
#> # Mean relative difference: 1
#> # --- fn(3)
#> # +++ c(1, 4, 3)
#> # "x"
#> # "1" 1
#> # "2" [-2-]{+4+}
#> # "3" 3
  # not ok - ut_cmp_equal(fn(3), c(1, 4, 3), filter = write.table)
  # # Test returned non-TRUE value:
  # # Mean relative difference: 1
  # # --- fn(3)
  # # +++ c(1, 4, 3)
  # # "x"
  # # "1" 1
  # # "2" [-2-]{+4+}
  # # "3" 3

  ## With ut_cmp_equal, an integer 1 is the same as a numeric 1
  ok(ut_cmp_equal(as.numeric(1), as.integer(1)))
#> ok - ut_cmp_equal(as.numeric(1), as.integer(1))
  # ok - ut_cmp_equal(as.numeric(1), as.integer(1))

  ## With ut_cmp_identical, they're not
  ok(ut_cmp_identical(as.numeric(1), as.integer(1)))
#> not ok - ut_cmp_identical(as.numeric(1), as.integer(1))
#> # Test returned non-TRUE value:
#> # --- as.numeric(1)
#> # +++ as.integer(1)
#> #  [-num-]{+int+} 1
  # not ok - ut_cmp_identical(as.numeric(1), as.integer(1))
  # # Test returned non-TRUE value:
  # # --- as.numeric(1)
  # # +++ as.integer(1)
  # #  [-num-]{+int+} 1

  ## all.equal() takes a tolerance parameter, for example:
  all.equal(0.01, 0.02, tolerance = 0.1)
#> [1] TRUE
  # [1] TRUE

  ## ...we can also give this to to ut_cmp_equal if we want a very
  ## approximate comparison
  ok(ut_cmp_equal(0.01, 0.02, tolerance = 0.1))
#> ok - ut_cmp_equal(0.01, 0.02, tolerance = 0.1)
  # ok - ut_cmp_equal(0.01, 0.02, tolerance = 0.1)

  ## We can make a comparison function of our own, and use
  ## deparse_depth to show the right expression in diff output
  cmp_noorder <- function (a, b) {
      sortlist <- function (x) if (length(x) > 0) x[order(names(x))] else x
      ut_cmp_identical(sortlist(a), sortlist(b), deparse_depth = -2)
  }
  ok(cmp_noorder(list(a=1, b=2), list(b=2, a=3)))
#> not ok - cmp_noorder(list(a = 1, b = 2), list(b = 2, a = 3))
#> # Test resulted in error:
#> #  unused argument (deparse_depth = -2)
#> # Whilst evaluating:
#> #  ut_cmp_identical(sortlist(a), sortlist(b), deparse_depth = -2)
#> # Stacktrace:
#> # -> .f(.x[[i]], ...)
#> # -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env, 
#> #     run_dont_run = run_dont_run), error = function(err) {
#> #     msg <- c(paste0("Failed to parse Rd in ", topic$file_in), 
#> #         i = err$message)
#> #     abort(msg, parent = err)
#> # })
#> # -> data_reference_topic(topic, pkg, examples_env = examples_env, 
#> #     run_dont_run = run_dont_run)
#> # -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env), 
#> #     topic = tools::file_path_sans_ext(topic$file_in), run_dont_run = run_dont_run)
#> # -> highlight_examples(code, topic, env = env)
#> # -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, 
#> #     env = child_env(env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> # -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> # -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, 
#> #     debug = debug, last = i == length(out), use_try = stop_on_error != 
#> #         2L, keep_warning = keep_warning, keep_message = keep_message, 
#> #     output_handler = output_handler, include_timing = include_timing)
#> # -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr, 
#> #     envir, enclos, user_handlers)), warning = wHandler, error = eHandler, 
#> #     message = mHandler)))
#> # -> handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr, 
#> #     envir, enclos, user_handlers)), warning = wHandler, error = eHandler, 
#> #     message = mHandler))
#> # -> try(f, silent = TRUE)
#> # -> tryCatch(expr, error = function(e) {
#> #     call <- conditionCall(e)
#> #     if (!is.null(call)) {
#> #         if (identical(call[[1L]], quote(doTryCatch))) 
#> #             call <- sys.call(-4L)
#> #         dcall <- deparse(call, nlines = 1L)
#> #         prefix <- paste("Error in", dcall, ": ")
#> #         LONG <- 75L
#> #         sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#> #         w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#> #         if (is.na(w)) 
#> #             w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], 
#> #                 type = "b")
#> #         if (w > LONG) 
#> #             prefix <- paste0(prefix, "\n  ")
#> #     }
#> #     else prefix <- "Error : "
#> #     msg <- paste0(prefix, conditionMessage(e), "\n")
#> #     .Internal(seterrmessage(msg[1L]))
#> #     if (!silent && isTRUE(getOption("show.error.messages"))) {
#> #         cat(msg, file = outFile)
#> #         .Internal(printDeferredWarnings())
#> #     }
#> #     invisible(structure(msg, class = "try-error", condition = e))
#> # })
#> # -> tryCatchList(expr, classes, parentenv, handlers)
#> # -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> # -> doTryCatch(return(expr), name, parentenv, handler)
#> # -> withCallingHandlers(withVisible(eval_with_user_handlers(expr, 
#> #     envir, enclos, user_handlers)), warning = wHandler, error = eHandler, 
#> #     message = mHandler)
#> # -> withVisible(eval_with_user_handlers(expr, envir, enclos, user_handlers))
#> # -> eval_with_user_handlers(expr, envir, enclos, user_handlers)
#> # -> eval(expr, envir, enclos)
#> # -> eval(expr, envir, enclos)
#> # -> ok(cmp_noorder(list(a = 1, b = 2), list(b = 2, a = 3)))
#> # -> tryCatch(withCallingHandlers(test, error = capture_calls), error = function(e) e)
#> # -> tryCatchList(expr, classes, parentenv, handlers)
#> # -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> # -> doTryCatch(return(expr), name, parentenv, handler)
#> # -> withCallingHandlers(test, error = capture_calls)
#> # -> cmp_noorder(list(a = 1, b = 2), list(b = 2, a = 3))
  # not ok - cmp_noorder(list(a = 1, b = 2), list(b = 2, a = 3))
  # # Test returned non-TRUE value:
  # # --- list(a = 1, b = 2)
  # # +++ list(b = 2, a = 3)
  # # $a
  # # [1] [-1-]{+3+}
  # #
  # # $b
  # # [1] 2

  # \dontshow{
    # Clear unittest result log, so our unittest failues don't fail example-building
    rm('outcomes', pos = unittest:::pkg_vars)
    rm('errors', pos = unittest:::pkg_vars)
#> Warning: object 'errors' not found
  # }