Test for and compare errors generated by code

A helper to catch expected errors and ensure they match what is expected

ut_cmp_error(code, expected_regexp, ignore.case = FALSE, perl = FALSE, fixed = FALSE)

Arguments

code

Code expression to test, should generate an error

expected_regexp

Regular expression the error should match

ignore.case

Passed to grepl

perl

Passed to grepl

fixed

Passed to grepl

Value

Returns TRUE if exp generates an error and matches expected_regexp. Returns a string with expected and actual error if exp generates an error but does not match. Returns "No error returned" if exp does not generate an error.

Examples

  ok(ut_cmp_error({
    stop("Hammer time")
  }, "hammer", ignore.case = TRUE), "Returned a hammer-based error")
#> ok - Returned a hammer-based error

  # \dontshow{
    # Clear unittest result log, so our unittest failues don't fail example-building
    rm('outcomes', pos = unittest:::pkg_vars)
    rm('errors', pos = unittest:::pkg_vars)
#> Warning: object 'errors' not found
  # }